+++ plutarch-ledger-api/src/Plutarch/LedgerApi/AssocMap.hs
@@ -81,7 +81,6 @@ import Data.Bifunctor (bimap)
 import Data.Foldable (foldl')
 import Data.Proxy (Proxy (Proxy))
 import Data.Traversable (for)
-import Plutarch.Bool (PSBool (PSFalse, PSTrue), psfalse, pstrue)
 import Plutarch.Builtin (
   pasMap,
   pdataImpl,
@@ -91,7 +90,12 @@ import Plutarch.Builtin (
  )
 import Plutarch.Internal.Term (punsafeBuiltin)
 import Plutarch.Internal.Witness (witness)
-import Plutarch.LedgerApi.Utils (Mret)
+import Plutarch.LedgerApi.Utils (
+  Mret,
+  PSBool (PSFalse, PSTrue),
+  psfalse,
+  pstrue,
+ )
 import Plutarch.Lift (
   PConstantDecl,
   PConstantRepr,